Fix GetScrollDeltaForScrollbarPart for FractionalScrollOffset
The ScrollbarController::GetScrollDeltaForScrollbarPart function
currently returns int, where under FractionalScrollOffset it
should return float. As a matter of fact, its call site is already
expecting a float.

This CL makes that function return a float value, as a result,
two layout tests pass.

Bug: 1150389
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2566152
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#832535}

+++ cc/input/scrollbar_controller.h
@@ -232,8 +232,8 @@ class CC_EXPORT ScrollbarController {
   gfx::Rect GetRectForScrollbarPart(const ScrollbarPart scrollbar_part) const;
 
   LayerImpl* GetLayerHitByPoint(const gfx::PointF position_in_widget) const;
-  int GetScrollDeltaForScrollbarPart(const ScrollbarPart scrollbar_part,
-                                     const bool jump_key_modifier) const;
+  float GetScrollDeltaForScrollbarPart(const ScrollbarPart scrollbar_part,
+                                       const bool jump_key_modifier) const;
 
   // Makes position_in_widget relative to the scrollbar.
   gfx::PointF GetScrollbarRelativePosition(const gfx::PointF position_in_widget,
@@ -257,17 +257,17 @@ class CC_EXPORT ScrollbarController {
                                     bool jump_key_modifier) const;
 
   // Calculates the delta based on position_in_widget and drag_origin.
-  int GetScrollDeltaForDragPosition(
+  float GetScrollDeltaForDragPosition(
       const gfx::PointF pointer_position_in_widget) const;
 
   // Returns the ratio of the scroller length to the scrollbar length. This is
   // needed to scale the scroll delta for thumb drag.
   float GetScrollerToScrollbarRatio() const;
 
-  int GetViewportLength() const;
+  float GetViewportLength() const;
 
   // Returns the pixel delta for a percent-based scroll of the scrollbar
-  int GetScrollDeltaForPercentBasedScroll() const;
+  float GetScrollDeltaForPercentBasedScroll() const;
 
   // Returns the page scale factor (i.e. pinch zoom factor). This is relevant
   // for root viewport scrollbar scrolling.
